[REFACT]: 인터페이스 개방 수정
