+++ Projects/App/Sources/AppContainer.swift
@@ -34,9 +34,7 @@ final class AppContainer:
   SignUpDependency,
   HomeDependency,
   SearchChallengeDependency,
-  MyPageDependency,
-  SettingDependency,
-  ProfileEditDependency {
+  MyPageDependency {
   func coordinator() -> Coordinating {
     let home = HomeContainer(dependency: self)
     let searchChallenge = SearchChallengeContainer(dependency: self)
@@ -58,14 +56,6 @@ final class AppContainer:
     return SignUpContainer(dependency: self)
   }()
   
-  lazy var profileEditContainable: ProfileEditContainable = {
-    return ProfileEditContainer(dependency: self)
-  }()
-
-  lazy var settingContainable: SettingContainable = {
-    return SettingContainer(dependency: self)
-  }()
-  
   // MARK: - UseCase
   lazy var logInUseCase: LogInUseCase = {
     return LogInUseCaseImpl(repository: logInRepository)
@@ -78,7 +68,7 @@ final class AppContainer:
   lazy var profileEditUseCase: ProfileEditUseCase = {
     return ProfileEditUseCaseImpl(repository: profileEditRepository)
   }()
-  
+
   // MARK: - Repository
   lazy var logInRepository: LogInRepository = {
     return LogInRepositoryImpl(dataMapper: LogInDataMapperImpl())
