+++ Projects/Presentation/MyPage/Implementations/Setting/SettingContainer.swift
@@ -7,26 +7,33 @@
 //
 
 import Core
-import MyPage
 import UseCase
 import ReportImpl
 
-public protocol SettingDependency: Dependency {
-  var profileEditContainable: ProfileEditContainable { get }
+protocol SettingDependency: Dependency {
+  var profileEditUseCase: ProfileEditUseCase { get }
 }
 
-public final class SettingContainer:
+protocol SettingContainable: Containable {
+  func coordinator(listener: SettingListener) -> Coordinating
+}
+
+final class SettingContainer:
   Container<SettingDependency>,
   SettingContainable,
+  ProfileEditDependency,
   ReportDependency {
+  var profileEditUseCase: ProfileEditUseCase { dependency.profileEditUseCase }
+  
   public func coordinator(listener: SettingListener) -> Coordinating {
     let viewModel = SettingViewModel()
     
+    let profileEdit = ProfileEditContainer(dependency: self)
     let report = ReportContainer(dependency: self)
     
     let coordinator = SettingCoordinator(
       viewModel: viewModel,
-      profileEditContainable: dependency.profileEditContainable,
+      profileEditContainable: profileEdit,
       reportContainable: report
     )
     coordinator.listener = listener
